Fixed closing
closeModal function can now be closed without sending any params. not event id,
because when closing it will always be the last modal.

ID was a good idea, but been deleted because after some common sense, we'll always close the last modal that been opened. so this.modals.pop should be enough.