Started binded styling
added to modal-service open function a "styling" option.
might change if a better styling idea will pop up.
Main idea now is to give the user an option to send an object of: "Styling Params"
including "height, width, minHeight, minWidth" (will be more)
and then to bind this object to the modal giving css.bind for inline-style using bindings.